requirements.txt
flask==1.1.2
requests==2.25.1
Pillow==8.2.0